requirements.txt
Django~=5.1.2
Pillow
django-ninja~=1.3.0
django-filter~=24.3